requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
scrapy
idna==3.4
Pillow==10.0.1
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6